feature/back/AMS-2

+++ ams_backend/src/main/java/com/gate_software/ams_backend/repository/CheckInRepository.java
@@ -3,5 +3,5 @@ package com.gate_software.ams_backend.repository;
 import com.gate_software.ams_backend.entity.CheckInRecords;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CheckInRepository extends JpaRepository<CheckInRecords, Long> {
+public interface CheckInRepository extends JpaRepository<CheckInRecords, Integer> {
 }
