+++ ams_backend/src/main/java/com/gate_software/ams_backend/repository/JobRepository.java
@@ -3,5 +3,5 @@ package com.gate_software.ams_backend.repository;
 import com.gate_software.ams_backend.entity.Job;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobRepository extends JpaRepository<Job, Long> {
+public interface JobRepository extends JpaRepository<Job, Integer> {
 }
